fix(astrascans): switch template and update url
