--- src/rust/madara/sources/astrascans/res/filters.json
@@ -1,305 +1,0 @@
-[
-	{
-		"type": "title"
-	},
-	{
-		"type": "group",
-		"name": "Status",
-		"filters": [
-			{
-				"type": "check",
-				"name": "Ongoing"
-			},
-			{
-				"type": "check",
-				"name": "On Hold"
-			},
-			{
-				"type": "check",
-				"name": "Cancelled"
-			},
-			{
-				"type": "check",
-				"name": "Completed"
-			},
-			{
-				"type": "check",
-				"name": "Upcoming"
-			}
-		]
-	},
-	{
-		"type": "select",
-		"name": "Genre Condition",
-		"options": [
-			"OR (having one of the selected genre)",
-			"AND (having all of the selected genres)"
-		]
-	},
-	{
-		"type": "group",
-		"name": "Genres",
-		"filters": [
-			{
-				"type": "genre",
-				"name": "Action",
-				"id": "action"
-			},
-			{
-				"type": "genre",
-				"name": "Adult",
-				"id": "adult"
-			},
-			{
-				"type": "genre",
-				"name": "Adventure",
-				"id": "adventure"
-			},
-			{
-				"type": "genre",
-				"name": "Anime",
-				"id": "anime"
-			},
-			{
-				"type": "genre",
-				"name": "Cartoon",
-				"id": "cartoon"
-			},
-			{
-				"type": "genre",
-				"name": "Comedy",
-				"id": "comedy"
-			},
-			{
-				"type": "genre",
-				"name": "Comic",
-				"id": "comic"
-			},
-			{
-				"type": "genre",
-				"name": "Cooking",
-				"id": "cooking"
-			},
-			{
-				"type": "genre",
-				"name": "Detective",
-				"id": "detective"
-			},
-			{
-				"type": "genre",
-				"name": "Doujinshi",
-				"id": "doujinshi"
-			},
-			{
-				"type": "genre",
-				"name": "Drama",
-				"id": "drama"
-			},
-			{
-				"type": "genre",
-				"name": "Ecchi",
-				"id": "ecchi"
-			},
-			{
-				"type": "genre",
-				"name": "Fantasy",
-				"id": "fantasy"
-			},
-			{
-				"type": "genre",
-				"name": "Full Color",
-				"id": "full-color"
-			},
-			{
-				"type": "genre",
-				"name": "Gang",
-				"id": "gang"
-			},
-			{
-				"type": "genre",
-				"name": "Gender Bender",
-				"id": "gender-bender"
-			},
-			{
-				"type": "genre",
-				"name": "Harem",
-				"id": "harem"
-			},
-			{
-				"type": "genre",
-				"name": "Historical",
-				"id": "historical"
-			},
-			{
-				"type": "genre",
-				"name": "Horror",
-				"id": "horror"
-			},
-			{
-				"type": "genre",
-				"name": "Josei",
-				"id": "josei"
-			},
-			{
-				"type": "genre",
-				"name": "Live action",
-				"id": "live-action"
-			},
-			{
-				"type": "genre",
-				"name": "Manga",
-				"id": "manga"
-			},
-			{
-				"type": "genre",
-				"name": "Manhua",
-				"id": "manhua"
-			},
-			{
-				"type": "genre",
-				"name": "Manhwa",
-				"id": "manhwa"
-			},
-			{
-				"type": "genre",
-				"name": "Martial Arts",
-				"id": "martial-arts"
-			},
-			{
-				"type": "genre",
-				"name": "Mature",
-				"id": "mature"
-			},
-			{
-				"type": "genre",
-				"name": "Mecha",
-				"id": "mecha"
-			},
-			{
-				"type": "genre",
-				"name": "Murim",
-				"id": "murim"
-			},
-			{
-				"type": "genre",
-				"name": "Mystery",
-				"id": "mystery"
-			},
-			{
-				"type": "genre",
-				"name": "One shot",
-				"id": "one-shot"
-			},
-			{
-				"type": "genre",
-				"name": "Otherworld",
-				"id": "otherworld"
-			},
-			{
-				"type": "genre",
-				"name": "Psychological",
-				"id": "psychological"
-			},
-			{
-				"type": "genre",
-				"name": "Reincarnation",
-				"id": "reincarnation"
-			},
-			{
-				"type": "genre",
-				"name": "Romance",
-				"id": "romance"
-			},
-			{
-				"type": "genre",
-				"name": "School Life",
-				"id": "school-life"
-			},
-			{
-				"type": "genre",
-				"name": "Sci-fi",
-				"id": "sci-fi"
-			},
-			{
-				"type": "genre",
-				"name": "Seinen",
-				"id": "seinen"
-			},
-			{
-				"type": "genre",
-				"name": "Shoujo",
-				"id": "shoujo"
-			},
-			{
-				"type": "genre",
-				"name": "Shoujo Ai",
-				"id": "shoujo-ai"
-			},
-			{
-				"type": "genre",
-				"name": "Shounen",
-				"id": "shounen"
-			},
-			{
-				"type": "genre",
-				"name": "Shounen Ai",
-				"id": "shounen-ai"
-			},
-			{
-				"type": "genre",
-				"name": "Slice of Life",
-				"id": "slice-of-life"
-			},
-			{
-				"type": "genre",
-				"name": "Smut",
-				"id": "smut"
-			},
-			{
-				"type": "genre",
-				"name": "Soft Yaoi",
-				"id": "soft-yaoi"
-			},
-			{
-				"type": "genre",
-				"name": "Soft Yuri",
-				"id": "soft-yuri"
-			},
-			{
-				"type": "genre",
-				"name": "Sports",
-				"id": "sports"
-			},
-			{
-				"type": "genre",
-				"name": "Supernatural",
-				"id": "supernatural"
-			},
-			{
-				"type": "genre",
-				"name": "Thriller",
-				"id": "thriller"
-			},
-			{
-				"type": "genre",
-				"name": "Tragedy",
-				"id": "tragedy"
-			},
-			{
-				"type": "genre",
-				"name": "Webtoon",
-				"id": "webtoon"
-			},
-			{
-				"type": "genre",
-				"name": "Yaoi",
-				"id": "yaoi"
-			},
-			{
-				"type": "genre",
-				"name": "Yuri",
-				"id": "yuri"
-			}
-		]
-	}
-]
