+++ src/rust/mangastream/sources/astrascans/Cargo.toml
@@ -9,4 +9,4 @@ crate-type = ["cdylib"]
 
 [dependencies]
 aidoku = { git = "https://github.com/Aidoku/aidoku-rs/" }
-madara_template = { path = "../../template" }
+mangastream_template = { path = "../../template" }
