+++ src/rust/mangastream/sources/astrascans/res/Icon.png
