+++ src/rust/mangastream/sources/astrascans/res/source.json
@@ -3,8 +3,8 @@
 		"id": "en.astrascans",
 		"lang": "en",
 		"name": "Astra Scans",
-		"version": 1,
-		"url": "https://astrascans.com",
+		"version": 2,
+		"url": "https://astrascans.net",
 		"nsfw": 0
 	},
 	"listings": [
